NO-ISSUE: Nest application lifecycle and app console commands under "flightctl app"
Regroup the "stop"/"start"/"restart" application-lifecycle commands and the
"--app"/"--remote-type" console flags under a new "flightctl app" parent
command, per review feedback:

  flightctl app start/stop/restart (device/NAME | fleet/NAME) --name APP
  flightctl app console device/NAME --name APP --type serial|vnc

The plain device-shell console ("flightctl console device/NAME") is
unchanged. The application console implementation (serial/VNC bridging)
moves to a new app_console.go, adapted to the "app console" flags.
