+++ cmd/flightctl/main.go
@@ -43,9 +43,7 @@ func NewFlightCtlCommand() *cobra.Command {
 	cmd.AddCommand(cli.NewCmdCertificate())
 	cmd.AddCommand(cli.NewCmdDownload())
 	cmd.AddCommand(cli.NewCmdLogs())
-	cmd.AddCommand(cli.NewCmdStop())
-	cmd.AddCommand(cli.NewCmdStart())
-	cmd.AddCommand(cli.NewCmdRestart())
+	cmd.AddCommand(cli.NewCmdApp())
 
 	return cmd
 }
