+++ internal/cli/console_test.go
@@ -1,18 +1,8 @@
 package cli
 
 import (
-	"context"
-	"crypto/x509"
-	"encoding/pem"
-	"net"
-	"net/http"
-	"net/http/httptest"
-	"strings"
 	"testing"
-	"time"
 
-	"github.com/flightctl/flightctl/internal/client"
-	"github.com/gorilla/websocket"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -21,8 +11,6 @@ func TestConsoleOptions_Validate(t *testing.T) {
 	tests := []struct {
 		name        string
 		args        []string
-		appName     string
-		remoteType  string
 		tty         bool
 		noTTY       bool
 		wantErr     bool
@@ -64,41 +52,11 @@ func TestConsoleOptions_Validate(t *testing.T) {
 			wantErr:     true,
 			errContains: "only devices can be connected to a console",
 		},
-		{
-			name:       "When --app and --remote-type serial are provided it should succeed",
-			args:       []string{"device/mydevice"},
-			appName:    "myvm",
-			remoteType: "serial",
-			wantErr:    false,
-		},
-		{
-			name:       "When --app and --remote-type vnc are provided it should succeed",
-			args:       []string{"device/mydevice"},
-			appName:    "myvm",
-			remoteType: "vnc",
-			wantErr:    false,
-		},
-		{
-			name:        "When --app is set but --remote-type is missing it should return an error",
-			args:        []string{"device/mydevice"},
-			appName:     "myvm",
-			wantErr:     true,
-			errContains: "--remote-type is required when --app is set",
-		},
-		{
-			name:        "When --remote-type is set without --app it should return an error",
-			args:        []string{"device/mydevice"},
-			remoteType:  "serial",
-			wantErr:     true,
-			errContains: "--remote-type requires --app",
-		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			o := DefaultConsoleOptions()
-			o.appName = tt.appName
-			o.remoteType = tt.remoteType
 			o.tty = tt.tty
 			o.noTTY = tt.noTTY
 
@@ -113,326 +71,4 @@ func TestConsoleOptions_Validate(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestBuildAppConsoleURL(t *testing.T) {
-	tests := []struct {
-		name          string
-		consoleServer string
-		deviceName    string
-		appName       string
-		remoteType    string
-		wantScheme    string
-		wantPath      string
-		wantQuery     string
-	}{
-		{
-			name:          "When server uses https it should produce a wss URL",
-			consoleServer: "https://console.example.com",
-			deviceName:    "dev1",
-			appName:       "myvm",
-			remoteType:    "serial",
-			wantScheme:    "wss",
-			wantPath:      "/ws/v1/devices/dev1/applications/myvm/console",
-		},
-		{
-			name:          "When server uses http it should produce a ws URL",
-			consoleServer: "http://console.example.com",
-			deviceName:    "dev1",
-			appName:       "myvm",
-			remoteType:    "serial",
-			wantScheme:    "ws",
-			wantPath:      "/ws/v1/devices/dev1/applications/myvm/console",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			o := DefaultConsoleOptions()
-			o.remoteType = tt.remoteType
-
-			got, err := o.buildAppConsoleURL(tt.consoleServer, tt.deviceName, tt.appName)
-			require.NoError(t, err)
-			assert.True(t, strings.HasPrefix(got, tt.wantScheme+"://"), "expected scheme %s in %s", tt.wantScheme, got)
-			assert.Contains(t, got, tt.wantPath)
-			assert.Contains(t, got, "consoleType="+tt.remoteType)
-		})
-	}
-}
-
-func TestConnectAppViaWS_HTTPErrors(t *testing.T) {
-	tests := []struct {
-		name        string
-		statusCode  int
-		body        string
-		errContains string
-	}{
-		{
-			name:        "When server returns 403 it should report auth error",
-			statusCode:  http.StatusForbidden,
-			body:        "Viewer role is not permitted",
-			errContains: "403",
-		},
-		{
-			name:        "When server returns 409 it should report duplicate session error",
-			statusCode:  http.StatusConflict,
-			body:        "a serial console session is already active",
-			errContains: "409",
-		},
-		{
-			name:        "When server returns 504 it should report timeout error",
-			statusCode:  http.StatusGatewayTimeout,
-			body:        "timed out waiting for agent",
-			errContains: "504",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			srv := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				http.Error(w, tt.body, tt.statusCode)
-			}))
-			defer srv.Close()
-
-			cfg := &client.Config{
-				RemoteAccessService: &client.Service{
-					Server:             srv.URL,
-					InsecureSkipVerify: true,
-				},
-			}
-
-			o := DefaultConsoleOptions()
-			o.remoteType = "serial"
-
-			err := o.connectAppViaWS(t.Context(), cfg, "dev1", "myvm", "")
-			require.Error(t, err)
-			assert.Contains(t, err.Error(), tt.errContains)
-		})
-	}
-}
-
-func TestBuildTLSConfigForConsole(t *testing.T) {
-	caKeyPair := generateSelfSignedCert(t)
-
-	tests := []struct {
-		name        string
-		consoleSvc  client.Service
-		authInfo    client.AuthInfo
-		wantRootCAs bool
-		wantCerts   bool
-		wantErrCA   bool
-		wantErrCert bool
-		serverName  string
-		insecure    bool
-	}{
-		{
-			name:       "When no TLS data is provided it should return a config with no extra CAs or certs",
-			consoleSvc: client.Service{},
-		},
-		{
-			name:        "When valid CA data is provided it should populate RootCAs",
-			consoleSvc:  client.Service{CertificateAuthorityData: caKeyPair.certPEM},
-			wantRootCAs: true,
-		},
-		{
-			name:       "When invalid CA data is provided it should return an error",
-			consoleSvc: client.Service{CertificateAuthorityData: []byte("not-a-cert")},
-			wantErrCA:  true,
-		},
-		{
-			name: "When valid client cert is provided it should add the certificate",
-			authInfo: client.AuthInfo{
-				ClientCertificateData: caKeyPair.certPEM,
-				ClientKeyData:         caKeyPair.keyPEM,
-			},
-			wantCerts: true,
-		},
-		{
-			name: "When invalid client cert data is provided it should return an error",
-			authInfo: client.AuthInfo{
-				ClientCertificateData: []byte("bad-cert"),
-				ClientKeyData:         []byte("bad-key"),
-			},
-			wantErrCert: true,
-		},
-		{
-			name:       "When TLSServerName is set it should be present in the TLS config",
-			consoleSvc: client.Service{TLSServerName: "myserver.example.com"},
-			serverName: "myserver.example.com",
-		},
-		{
-			name:       "When InsecureSkipVerify is true it should be set in the TLS config",
-			consoleSvc: client.Service{InsecureSkipVerify: true},
-			insecure:   true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := buildTLSConfigForConsole(&tt.consoleSvc, tt.authInfo)
-			if tt.wantErrCA || tt.wantErrCert {
-				require.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			require.NotNil(t, got)
-			if tt.wantRootCAs {
-				assert.NotNil(t, got.RootCAs)
-			} else {
-				assert.Nil(t, got.RootCAs)
-			}
-			if tt.wantCerts {
-				assert.NotEmpty(t, got.Certificates)
-			} else {
-				assert.Empty(t, got.Certificates)
-			}
-			assert.Equal(t, tt.serverName, got.ServerName)
-			assert.Equal(t, tt.insecure, got.InsecureSkipVerify)
-		})
-	}
-}
-
-// certKeyPair holds PEM-encoded self-signed cert and key for testing.
-type certKeyPair struct {
-	certPEM []byte
-	keyPEM  []byte
-}
-
-// extractServerCert extracts the PEM-encoded server certificate from a TLS test server.
-func extractServerCert(t *testing.T, srv *httptest.Server) []byte {
-	t.Helper()
-	certDER := srv.TLS.Certificates[0].Certificate[0]
-	return pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: certDER})
-}
-
-// generateSelfSignedCert creates a minimal self-signed cert/key pair for test use by
-// extracting them from a short-lived httptest.TLSServer.
-func generateSelfSignedCert(t *testing.T) certKeyPair {
-	t.Helper()
-
-	srv := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {}))
-	defer srv.Close()
-
-	certDER := srv.TLS.Certificates[0].Certificate[0]
-	certPEM := pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: certDER})
-
-	privKey := srv.TLS.Certificates[0].PrivateKey
-	keyBytes, err := x509.MarshalPKCS8PrivateKey(privKey)
-	require.NoError(t, err)
-	keyPEM := pem.EncodeToMemory(&pem.Block{Type: "PRIVATE KEY", Bytes: keyBytes})
-
-	return certKeyPair{certPEM: certPEM, keyPEM: keyPEM}
-}
-
-func TestConnectVNCViaWS_MissingRemoteAccessService(t *testing.T) {
-	o := DefaultConsoleOptions()
-	o.remoteType = "vnc"
-
-	cfg := &client.Config{}
-	err := o.connectVNCViaWS(context.Background(), cfg, "dev1", "myvm", "")
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "remote access service is not configured")
-}
-
-func TestConnectVNCViaWS_HTTPError(t *testing.T) {
-	srv := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		http.Error(w, "forbidden", http.StatusForbidden)
-	}))
-	defer srv.Close()
-
-	o := DefaultConsoleOptions()
-	o.remoteType = "vnc"
-
-	cfg := &client.Config{
-		RemoteAccessService: &client.Service{
-			Server:                   srv.URL,
-			CertificateAuthorityData: extractServerCert(t, srv),
-		},
-	}
-	err := o.connectVNCViaWS(context.Background(), cfg, "dev1", "myvm", "")
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "403")
-}
-
-func TestBridgeVNCClient_ClientDisconnectDoesNotCloseWebSocket(t *testing.T) {
-	// Set up a WebSocket echo server that stays open for the duration of the test.
-	upgrader := websocket.Upgrader{CheckOrigin: func(_ *http.Request) bool { return true }}
-	wsServerDone := make(chan struct{})
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		defer close(wsServerDone)
-		conn, err := upgrader.Upgrade(w, r, nil)
-		if err != nil {
-			return
-		}
-		defer conn.Close()
-		for {
-			msgType, data, err := conn.ReadMessage()
-			if err != nil {
-				return
-			}
-			if err := conn.WriteMessage(msgType, data); err != nil {
-				return
-			}
-		}
-	}))
-	defer srv.Close()
-
-	wsURL := "ws" + strings.TrimPrefix(srv.URL, "http")
-	wsConn, _, err := websocket.DefaultDialer.Dial(wsURL, nil)
-	require.NoError(t, err)
-	defer wsConn.Close()
-
-	wsRecvCh := make(chan []byte, 64)
-
-	// Run the WebSocket reader goroutine.
-	go func() {
-		defer close(wsRecvCh)
-		for {
-			_, data, err := wsConn.ReadMessage()
-			if err != nil {
-				return
-			}
-			wsRecvCh <- data
-		}
-	}()
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	// First VNC client: connect, send data, disconnect.
-	serverSide, clientSide := net.Pipe()
-	go func() {
-		defer serverSide.Close()
-		_, _ = serverSide.Write([]byte("hello"))
-		// read the echo back
-		buf := make([]byte, 5)
-		_, _ = serverSide.Read(buf)
-		// disconnect
-	}()
-
-	o := DefaultConsoleOptions()
-	o.bridgeVNCClient(ctx, clientSide, wsConn, wsRecvCh)
-	clientSide.Close()
-
-	// WebSocket should still be open: send a second message.
-	require.NoError(t, wsConn.WriteMessage(websocket.BinaryMessage, []byte("still-open")))
-
-	// Confirm the echo server is still running (wsServerDone not closed yet).
-	select {
-	case <-wsServerDone:
-		t.Fatal("WebSocket server closed unexpectedly after first VNC client disconnected")
-	case <-time.After(100 * time.Millisecond):
-		// expected: server still running
-	}
-}
-
-func TestConnectAppViaWS_MissingRemoteAccessService(t *testing.T) {
-	o := DefaultConsoleOptions()
-	o.remoteType = "serial"
-
-	cfg := &client.Config{}
-
-	err := o.connectAppViaWS(t.Context(), cfg, "dev1", "myvm", "")
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "remote access service is not configured")
 }
